# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StockFlow — stock management, storefront, social console
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      parameters:
        - in: query
          name: variantId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
            enum: [in, out, transfer, adjust]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovementWithRefs"
    post:
      operationId: createMovement
      tags: [movements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementWithRefs"
  /movements/scan:
    post:
      operationId: scanBarcode
      tags: [movements]
      summary: Resolve a barcode/SKU and (optionally) post a movement in the same call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
  /orders:
    get:
      operationId: listPurchaseOrders
      tags: [orders]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrderWithRefs"
    post:
      operationId: createPurchaseOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithRefs"
  /orders/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderWithRefs"
...
  /printers:
    get:
      operationId: listPrinters
      tags: [printers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Printer"
    post:
      operationId: createPrinter
      tags: [printers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrinterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Printer"
  /printers/{id}:
    delete:
      operationId: deletePrinter
      tags: [printers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    delete:
      operationId: deleteBanner
      tags: [banners]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt]
    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        zone:
          type: ["string", "null"]
        aisle:
          type: ["string", "null"]
        shelf:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, zone, aisle, shelf, createdAt]
    CreateLocationInput:
      type: object
      properties:
        name:
          type: string
        zone:
          type: ["string", "null"]
        aisle:
          type: ["string", "null"]
        shelf:
          type: ["string", "null"]
      required: [name]
    Variant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        threshold:
          type: integer
        onHand:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, productId, sku, barcode, size, color, threshold, onHand, createdAt]
...
    ProductWithDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        reference:
          type: string
        description:
          type: ["string", "null"]
        priceHt:
          type: number
        vatRate:
          type: number
        priceTtc:
          type: number
        imageUrl:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        visibleOnStorefront:
          type: boolean
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        totalOnHand:
          type: integer
      required:
        - id
        - name
        - reference
        - description
        - priceHt
        - vatRate
        - priceTtc
        - imageUrl
        - gallery
        - categoryId
        - categoryName
        - supplierId
        - supplierName
        - visibleOnStorefront
        - updatedBy
        - updatedAt
        - createdAt
        - variants
        - totalOnHand
...
    MovementWithRefs:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, transfer, adjust]
        variantId:
          type: integer
        sku:
          type: string
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
        fromLocationId:
          type: ["integer", "null"]
        fromLocationName:
          type: ["string", "null"]
        toLocationId:
          type: ["integer", "null"]
        toLocationName:
          type: ["string", "null"]
        operator:
          type: string
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - variantId
        - sku
        - productId
        - productName
        - productImageUrl
        - quantity
        - fromLocationId
        - fromLocationName
        - toLocationId
        - toLocationName
        - operator
        - reason
        - createdAt
    CreateMovementInput:
      type: object
      properties:
        variantId:
          type: integer
        type:
          type: string
          enum: [in, out, transfer, adjust]
        quantity:
          type: integer
        fromLocationId:
          type: ["integer", "null"]
        toLocationId:
          type: ["integer", "null"]
        operator:
          type: string
        reason:
          type: ["string", "null"]
      required: [variantId, type, quantity, operator]
    ScanInput:
      type: object
      properties:
        code:
          type: string
        mode:
          type: string
          enum: [lookup, in, out, adjust]
        quantity:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        operator:
          type: ["string", "null"]
      required: [code, mode]
    ScanResult:
      type: object
      properties:
        matched:
          type: boolean
        variant:
          anyOf:
            - $ref: "#/components/schemas/Variant"
            - type: "null"
        product:
          anyOf:
            - $ref: "#/components/schemas/ProductWithDetails"
            - type: "null"
        movement:
          anyOf:
            - $ref: "#/components/schemas/MovementWithRefs"
            - type: "null"
        message:
          type: string
      required: [matched, variant, product, movement, message]
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        variantId:
          type: integer
        sku:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPriceHt:
          type: number
      required: [id, variantId, sku, productName, quantity, unitPriceHt]
    PurchaseOrderWithRefs:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
          enum: [draft, sent, received, cancelled]
        totalHt:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        expectedAt:
          type: ["string", "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
      required:
        - id
        - reference
        - supplierId
        - supplierName
        - status
        - totalHt
        - notes
        - createdAt
        - expectedAt
        - items
    CreatePurchaseOrderItemInput:
      type: object
      properties:
        variantId:
          type: integer
        quantity:
          type: integer
        unitPriceHt:
          type: number
      required: [variantId, quantity, unitPriceHt]
    CreatePurchaseOrderInput:
      type: object
      properties:
        supplierId:
          type: integer
        notes:
          type: ["string", "null"]
        expectedAt:
          type: ["string", "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePurchaseOrderItemInput"
      required: [supplierId, items]
    UpdatePurchaseOrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, received, cancelled]
      required: [status]
...
    Printer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        connection:
          type: string
          enum: [network, bluetooth]
        address:
          type: string
        width:
          type: string
          enum: ["58mm", "80mm"]
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, connection, address, width, isDefault, createdAt]
    CreatePrinterInput:
      type: object
      properties:
        name:
          type: string
        connection:
          type: string
          enum: [network, bluetooth]
        address:
          type: string
        width:
          type: string
          enum: ["58mm", "80mm"]
        isDefault:
          type: boolean
      required: [name, connection, address, width]
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        startsAt:
          type: ["string", "null"]
          format: date-time
        endsAt:
          type: ["string", "null"]
          format: date-time
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, subtitle, imageUrl, linkUrl, startsAt, endsAt, active, createdAt]
    CreateBannerInput:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        startsAt:
          type: ["string", "null"]
          format: date-time
        endsAt:
          type: ["string", "null"]
          format: date-time
        active:
          type: boolean
      required: [title, imageUrl]
...